Enable bilock bench and tests
